Feature: 对齐 Matrix 模板维度类型并补充初始化测试
1. 将 Matrix 矩阵库相关行列和维度模板参数由 size_t 调整为 int，保持与 Eigen 固定尺寸模板风格一致。
2. 为矩阵、向量、切片、方阵、稀疏向量和 Dual 等模板补充非负维度 static_assert，提前拦截非法尺寸。
3. 增加 Matrix initializer_list 构造与赋值能力，并补充 MatrixTest 覆盖初始化和 isApprox 行为。
4. 统一格式化 Matrix 目录相关实现，整理矩阵库代码风格。

+++ Matrix/Vector4.hpp
@@ -43,87 +43,54 @@
 
 #include "Vector.hpp"
 
-namespace matrix
-{
+namespace matrix {
 
-template<typename Type>
-class Vector4 : public Vector<Type, 4>
-{
-public:
-	using Matrix41 = Matrix<Type, 4, 1>;
+template <typename Type>
+class Vector4 : public Vector<Type, 4> {
+ public:
+  using Matrix41 = Matrix<Type, 4, 1>;
 
-	Vector4() = default;
+  Vector4() = default;
 
-	Vector4(const Matrix41 &other) :
-		Vector<Type, 4>(other)
-	{
-	}
+  Vector4(const Matrix41& other) : Vector<Type, 4>(other) {}
 
-	explicit Vector4(const Type data_[3]) :
-		Vector<Type, 4>(data_)
-	{
-	}
+  explicit Vector4(const Type data_[3]) : Vector<Type, 4>(data_) {}
 
-	Vector4(Type x1, Type x2, Type x3, Type x4)
-	{
-		Vector4 &v(*this);
-		v(0) = x1;
-		v(1) = x2;
-		v(2) = x3;
-		v(3) = x4;
-	}
+  Vector4(Type x1, Type x2, Type x3, Type x4) {
+    Vector4& v(*this);
+    v(0) = x1;
+    v(1) = x2;
+    v(2) = x3;
+    v(3) = x4;
+  }
 
-	template<size_t P, size_t Q>
-	Vector4(const Slice<Type, 4, 1, P, Q> &slice_in) : Vector<Type, 4>(slice_in)
-	{
-	}
+  template <int P, int Q>
+  Vector4(const Slice<Type, 4, 1, P, Q>& slice_in)
+      : Vector<Type, 4>(slice_in) {}
 
-	template<size_t P, size_t Q>
-	Vector4(const Slice<Type, 1, 4, P, Q> &slice_in) : Vector<Type, 4>(slice_in)
-	{
-	}
+  template <int P, int Q>
+  Vector4(const Slice<Type, 1, 4, P, Q>& slice_in)
+      : Vector<Type, 4>(slice_in) {}
 
-	/**
-	 * Override matrix ops so Vector4 type is returned
-	 */
+  /**
+   * Override matrix ops so Vector4 type is returned
+   */
 
-	Vector4 operator+(Vector4 other) const
-	{
-		return Matrix41::operator+(other);
-	}
+  Vector4 operator+(Vector4 other) const { return Matrix41::operator+(other); }
 
-	Vector4 operator+(Type scalar) const
-	{
-		return Matrix41::operator+(scalar);
-	}
+  Vector4 operator+(Type scalar) const { return Matrix41::operator+(scalar); }
 
-	Vector4 operator-(Vector4 other) const
-	{
-		return Matrix41::operator-(other);
-	}
+  Vector4 operator-(Vector4 other) const { return Matrix41::operator-(other); }
 
-	Vector4 operator-(Type scalar) const
-	{
-		return Matrix41::operator-(scalar);
-	}
+  Vector4 operator-(Type scalar) const { return Matrix41::operator-(scalar); }
 
-	Vector4 operator-() const
-	{
-		return Matrix41::operator-();
-	}
+  Vector4 operator-() const { return Matrix41::operator-(); }
 
-	Vector4 operator*(Type scalar) const
-	{
-		return Matrix41::operator*(scalar);
-	}
+  Vector4 operator*(Type scalar) const { return Matrix41::operator*(scalar); }
 
-	Type operator*(Vector4 b) const
-	{
-		return Vector<Type, 4>::operator*(b);
-	}
-
+  Type operator*(Vector4 b) const { return Vector<Type, 4>::operator*(b); }
 };
 
 using Vector4f = Vector4<float>;
 
-} // namespace matrix
+}  // namespace matrix
